refactor: Extract datasource result filters to separate file
