+++ lib/modules/datasource/conan/index.ts
@@ -3,9 +3,9 @@ import { load } from 'js-yaml';
 import { logger } from '../../../logger';
 import { cache } from '../../../util/cache/package/decorator';
 import { GithubHttp } from '../../../util/http/github';
+import type { HttpResponse } from '../../../util/http/types';
 import { ensureTrailingSlash, joinUrlParts } from '../../../util/url';
 import * as allVersioning from '../../versioning';
-import { isArtifactoryServer } from '../common';
 import { Datasource } from '../datasource';
 import type {
   DigestConfig,
@@ -26,6 +26,12 @@ import type {
   ConanRevisionsJSON,
   ConanYAML,
 } from './types';
+
+function isArtifactoryServer<T = unknown>(
+  res: HttpResponse<T> | undefined
+): boolean {
+  return is.string(res?.headers['x-jfrog-version']);
+}
 
 export class ConanDatasource extends Datasource {
   static readonly id = datasource;
