+++ lib/workers/repository/process/vulnerabilities.ts
@@ -6,7 +6,6 @@ import { parseCvssVector } from 'vuln-vects';
 import { getManagerConfig, mergeChildConfig } from '../../../config';
 import type { PackageRule, RenovateConfig } from '../../../config/types';
 import { logger } from '../../../logger';
-import { getDefaultVersioning } from '../../../modules/datasource';
 import type {
   PackageDependency,
   PackageFile,
@@ -24,6 +23,7 @@ import type {
   SeverityDetails,
   Vulnerability,
 } from './types';
+import { getDefaultVersioning } from '../../../modules/datasource/common';
 
 export class Vulnerabilities {
   private osvOffline: OsvOffline | undefined;
